==> 2.creating_data_for_transformers.ipynb <==
import pandas as pd
import csv
import numpy as np
import matplotlib.pyplot as plt
import datetime 
import seaborn as sns

from google.colab import drive
drive.mount('/content/drive')

cd "/content/drive/Shareddrives/Machine Learning - Transformers"

data =  pd.read_csv('./Data/final_new.csv')
data 

stations = data.station_id.unique()

data['time'] = pd.to_datetime(data['time'])

data.isna().sum()

data = data.set_index(["time", "station_id"])

time = pd.date_range(start='2014-05-01 00:00:00', end='2015-04-30 23:00:00', freq='H')
time

data1 = data.reset_index()

data_new = pd.DataFrame(index=pd.MultiIndex.from_product([time, stations], names=['time', 'station_id'])).reset_index()

data_new

data_final = pd.merge(data_new, data, how='left', left_on=['time', 'station_id'], right_on = ['time', 'station_id'])

data_final

data_final['wind_speed'] = data_final['wind_speed'].interpolate(method="quadratic").ffill().bfill()
data_final['wind_direction'] = data_final['wind_direction'].interpolate(method="nearest").ffill().bfill()
data_final['humidity'] = data_final['humidity'].interpolate(method="quadratic").ffill().bfill()
data_final['pressure'] = data_final['pressure'].interpolate(method="quadratic").ffill().bfill()
data_final['weather'] = data_final['weather'].interpolate(method='nearest').ffill().bfill()
data_final['temperature'] = data_final['temperature'].interpolate(method="quadratic").ffill().bfill()
data_final['PM25_Concentration'] = data_final['temperature'].interpolate(method="quadratic").ffill().bfill()

data_final.drop('district_id', axis=1, inplace=True)

lat_dic = {}
lon_dic = {}
for i in data1[data1['time']=='2014-05-01 00:00:00'].iterrows():
  lat_dic[i[1]['station_id']] = i[1]['latitude']
  lon_dic[i[1]['station_id']] = i[1]['longitude']

for i in stations:
  data_final.loc[data_final['station_id']==i, 'latitude'] = lat_dic[i]
  data_final.loc[data_final['station_id']==i, 'longitude'] = lon_dic[i]

data_final

data_final.to_csv('Data/data_final_transformer.csv', index=False)
