==> 3.prepare_data_and_folds.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import csv
from sklearn.neighbors import KNeighborsRegressor
import time
import datetime
from sklearn.model_selection import train_test_split, KFold, GridSearchCV
from sklearn.ensemble import RandomForestRegressor
from sklearn.datasets import make_regression
from sklearn.linear_model import LinearRegression
from sklearn.svm import SVR
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.datasets import make_hastie_10_2
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.metrics import mean_squared_error

from google.colab import drive 
drive.mount('/content/drive')

cd /content/drive/Shareddrives/Machine Learning - Transformers/

def oneHotEncode(df, cols):
  for i in cols:
    x = pd.get_dummies(df.loc[:,i], prefix=i)
    df.drop(i, axis=1, inplace=True)
    df = pd.concat([df,x], axis=1)
  return df

cols = ['weather', 'wind_direction']

# final_new_interpolated = pd.read_csv('Data/data_final_transformer.csv')
final_new_interpolated = pd.read_csv('Data/final_new_Interpolated.csv')
final_new_interpolated = oneHotEncode(final_new_interpolated, cols)
final_new_interpolated

final_mm = pd.read_csv('Data/final_mm.csv')
final_mm = oneHotEncode(final_mm, cols)
# final_mm
final_std = pd.read_csv('Data/final_std.csv')
final_std = oneHotEncode(final_std, cols)
# final_std

final_mm[final_mm["station_id"]==1035].head()

# LOOKING AT THE DATA

final_new_interpolated.head()

final_mm.head()

final_std.head()

# Creating months attribute

months = pd.Series( [ ( datetime.datetime.strptime(time,'%Y-%m-%d %H:%M:%S').month ) for time in final_new_interpolated['time'] ], copy = True )
months

# Randomly taking 24 stations as Train and 12 as test

final_new_interpolated.shape

final_months = []
final_months_std = []
final_months_mm = []
for month in range(1, 13):
  final_months.append(final_new_interpolated[months == month])
  final_months_mm.append(final_mm[months == month])
  final_months_std.append(final_std[months == month])
final_months[7]

month = 7

stations = np.unique(final_months[month]['station_id'])

splitter = KFold(n_splits=3, shuffle=True, random_state=1234)

for fold, (train, test) in enumerate(splitter.split(stations)):
    train_stations = stations[train]
    test_stations = stations[test]
    train_data = final_months[month][final_months[month].station_id.isin(train_stations)]
    test_data = final_months[month][final_months[month].station_id.isin(test_stations)]

    train_data_mm = final_months_mm[month][final_months_mm[month].station_id.isin(train_stations)]
    test_data_mm = final_months_mm[month][final_months_mm[month].station_id.isin(test_stations)]

    train_data_std = final_months_std[month][final_months_std[month].station_id.isin(train_stations)]
    test_data_std = final_months_std[month][final_months_std[month].station_id.isin(test_stations)]
    
    # train_data.to_csv('Data/data_final_transformers_'+str(fold)+'_train.csv', index=None)
    # test_data.to_csv('Data/data_final_transformers_'+str(fold)+'_test.csv', index=None)
    train_data.to_csv('Data/interpolated/fold_'+str(fold)+'_train.csv', index=None)
    test_data.to_csv('Data/interpolated/fold_'+str(fold)+'_test.csv', index=None)

    train_data_mm.to_csv('Data/interpolated/fold_'+str(fold)+'_train_mm.csv', index=None)
    test_data_mm.to_csv('Data/interpolated/fold_'+str(fold)+'_test_mm.csv', index=None)

    train_data_std.to_csv('Data/interpolated/fold_'+str(fold)+'_train_std.csv', index=None)
    test_data_std.to_csv('Data/interpolated/fold_'+str(fold)+'_test_std.csv', index=None)
